nivo/bar chart dependencies installed
